fix issue with behaviour of move_over_item
The issue occurred under these conditions:

1: The setup procedure contains a move_over_item call
2: The source of the move_over_item call does not originally exist
3: The setup procedure deploys a file to the source of the move_over_item call
4: The setup procedure is invoked multiple times

These conditions would lead to a duplicate of the file deployed to the source of
the move_over_item call because the first call to move_over_item would not be
'remembered' and the second invocation would therefore move over the deployed
file before it would get deployed again. In the common case where the deployed
file would source the moved-over file, this would lead to a recursive source.

This change will make sure that a successful call to move_over_item is
'remembered' even if the source does not exist, thereby preventing the issue.

+++ scripts/setup.sh
@@ -211,14 +211,6 @@ move_over_item () {
 
     echo "Moving \"$source\" over to \"$destination\""
 
-    if ! [ -e "$source" ]; then
-        echo "Move-over source \"$source\" does not exist. Nothing to move over."
-        return 0
-    fi
-
-    ensure_outside "$source" "$REPO_PATH" || return
-    ensure_outside "$destination" "$REPO_PATH" || return
-
     if [ -e "$destination" ] || [ -L "$destination" ]; then
         local line
         while read line; do
@@ -227,13 +219,24 @@ move_over_item () {
                 return 0
             fi
         done < "$MOVE_OVER_FILES_PATH"
+    fi
 
+    ensure_outside "$source" "$REPO_PATH" || return
+    ensure_outside "$destination" "$REPO_PATH" || return
+
+    if [ -e "$destination" ] || [ -L "$destination" ]; then
         local destination_name=$(basename "$destination")
         backup_item "$destination" "$BACKUP_DIR/move-overs/$destination_name" || return
         rm "$destination"
     fi
 
     echo "$destination_canonical" >> "$MOVE_OVER_FILES_PATH"
+
+    if ! [ -e "$source" ]; then
+        echo "Move-over source \"$source\" does not exist. Nothing to move over."
+        return 0
+    fi
+
     ensure_containing_dir "$destination"
     mv -n -T "$source" "$destination"
     return
